requirements.txt
mlflow>=2.10.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
PyYAML>=6.0
pytest>=7.4.0
flake8>=6.0.0
